Extract Chiling workbench workflow views
