+++ web/chiling-workbench/app.js
@@ -1,21 +1,21 @@
 import {
   clampNumber,
   escapeHtml,
-  lineBreaks,
   normalizeSubtitleText,
-  formatRelativeTime,
 } from "./src/format.js";
-import {
-  defaultDeliverables,
-  taskStatusLabel,
-} from "./src/task-model.js";
+import {} from "./src/task-model.js";
 import { createInitialState } from "./src/state.js";
-import { phonePreview } from "./src/components/ui.js";
+import {} from "./src/components/ui.js";
 import { renderTopbar } from "./src/components/topbar.js";
 import { bindDelegatedClick, find, findAll } from "./src/dom.js";
 import { render as renderLoginView } from "./src/views/login.js";
 import { render as renderDashboardView } from "./src/views/dashboard.js";
 import { render as renderCreateView } from "./src/views/create.js";
+import { render as renderReviewView } from "./src/views/review.js";
+import { render as renderGeneratingView } from "./src/views/generating.js";
+import { render as renderDeliveryView } from "./src/views/delivery.js";
+import { render as renderAdminView } from "./src/views/admin.js";
+import { render as renderTaskDetailDrawerView } from "./src/views/detail-drawer.js";
 
 const appRoot = document.querySelector("#app");
 const toastRoot = document.querySelector("#toast-root");
@@ -531,450 +531,45 @@ function renderCreate() {
 }
 
 function renderReview() {
-  return shell(
-    `
-      <section class="page-head">
-        <div>
-          <h1 class="hero-title hero-title--compact">人工审核</h1>
-          <p class="lede">确认文案、字幕规则与授权状态，审核通过后才会进入生产。</p>
-        </div>
-        <button class="button" data-route="create">退回修改</button>
-      </section>
-
-      <section class="recent-card" style="margin-top: 28px;">
-        <div class="page-head" style="margin-bottom: 0;">
-          <div>
-            <h2 class="section-title">后台生产队列</h2>
-            <p class="lede">这里显示已提交任务的后台状态、下一步动作和是否等待人工确认。</p>
-          </div>
-          <button class="button button--small" data-refresh-queue>刷新队列</button>
-        </div>
-        <div class="delivery-list" style="margin-top: 24px;">
-          ${renderQueueRows()}
-        </div>
-      </section>
-
-      ${renderProductionServiceStatusPanel()}
-
-      <section class="recent-card" style="margin-top: 28px;">
-        <div class="page-head" style="margin-bottom: 0;">
-          <div>
-            <h2 class="section-title">生产执行队列</h2>
-            <p class="lede">这里只显示已完成二次确认、等待操作员执行的生产请求。</p>
-          </div>
-          <button class="button button--small" data-refresh-production-requests>刷新执行队列</button>
-        </div>
-        <div class="delivery-list" style="margin-top: 24px;">
-          ${renderProductionRequestRows()}
-        </div>
-      </section>
-
-      <section class="review-grid">
-        <aside class="review-card">
-          <span class="pill pill--blue">参考预览</span>
-          ${phonePreview(state.form.referencePreview, "参考视频", "phone--large")}
-          <div class="chip-row">
-            <span class="pill">${state.form.duration}s</span>
-            <span class="pill">${state.form.resolution}</span>
-          </div>
-        </aside>
-
-        <article class="review-card">
-          ${renderReviewDraftPanel()}
-          <h2 class="section-title">文案与字幕</h2>
-          <p class="lede">短句结尾不显示逗号、句号等标点。这里修改后会带入生产。</p>
-          <textarea class="review-script" data-field="script">${escapeHtml(state.form.script)}</textarea>
-          <div class="chip-row">
-            <span class="pill pill--blue">短句显示</span>
-            <span class="pill pill--green">句尾无标点</span>
-          </div>
-          <div class="form-actions">
-            <button class="button" data-clean-subtitles>清理句尾标点</button>
-            <button class="button" data-save-review>保存审核稿</button>
-            <button class="button button--primary" data-approve-review>审核通过</button>
-            <button class="button button--primary" data-start-generation>确认并生成</button>
-          </div>
-        </article>
-
-        <aside class="review-card">
-          <h2 class="section-title">审核详情侧栏</h2>
-          <p class="lede">从列表滑出，不打断主页面。</p>
-          <div class="check-list">
-            ${checkRow("素材授权", "已确认", "ok")}
-            ${checkRow("肖像授权", "已确认", "ok")}
-            ${checkRow("字幕规则", "已确认", "ok")}
-            ${checkRow("画面方向", "需确认", "warning")}
-          </div>
-          <div class="review-note">
-            <strong>当前生产设置</strong>
-            <span class="muted">单批 ${state.form.count} 条，单条 ${state.form.duration} 秒，清晰度 ${state.form.resolution}。</span>
-          </div>
-        </aside>
-      </section>
-    `,
-    "review",
-  );
+  return shell(renderReviewView({ state }), "review");
 }
 
 function renderGenerating() {
-  const task = state.currentTask;
-  const payload = task?.payload || state.form;
-  const taskStatus = !task ? "待创建" : task.status === "queued" ? "排队中" : "处理中";
-  const taskButtonClass = task ? "button button--primary button--loading" : "button button--primary";
-  const bannerTitle = task ? "已提交，生产任务正在进行" : "等待任务提交";
-  const bannerSubtitle = task ? "可以离开页面，完成后会在交付区提示。" : "请从创建作品或人工审核页提交任务。";
-  const visibleProgress = task ? state.progress : 0;
-
-  return shell(
-    `
-      <section class="generation-layout">
-        <article class="generation-main">
-          <div class="state-banner">
-            <i class="status-dot status-dot--green"></i>
-            <div>
-              <strong>${bannerTitle}</strong>
-              <span class="muted">${bannerSubtitle}</span>
-            </div>
-          </div>
-          <h1 class="hero-title hero-title--compact">生成中</h1>
-          <p class="lede">系统正在处理画面、字幕和交付包。当前任务 ${task?.id || "待创建"}，状态会持续更新。</p>
-          <div class="generation-body">
-            <div>
-              <div class="stage-list">
-                ${generationStages(task)}
-              </div>
-              <div class="form-actions" style="justify-content: flex-start;">
-                <button class="button" data-route="review">查看审核稿</button>
-                <button class="${taskButtonClass}" disabled>${taskStatus}</button>
-              </div>
-            </div>
-            <div class="progress-ring progress-ring--large" data-progress-ring data-progress="${visibleProgress}" style="--progress: ${visibleProgress};"></div>
-          </div>
-          ${renderOperationPanel()}
-          ${renderGenerationApprovalPanel()}
-          ${renderProductionPrepPanel()}
-        </article>
-
-        <aside class="generation-side">
-          <span class="pill pill--amber">任务抽屉</span>
-          <h2 class="section-title" style="margin-top: 24px;">当前批次</h2>
-          <p class="lede">生产 ${payload.count} 条，每条最长 ${payload.duration} 秒，清晰度 ${payload.resolution}。</p>
-          ${phonePreview(state.form.portraitPreview, "指定肖像", "phone--large")}
-        </aside>
-      </section>
-    `,
-    "review",
-  );
+  return shell(renderGeneratingView({ state }), "review");
 }
 
 function renderDelivery() {
-  const task = state.currentTask;
-  const payload = task?.payload || state.form;
-  const deliverables = state.deliverables.length ? state.deliverables : defaultDeliverables(payload);
-
-  return shell(
-    `
-      <section class="page-head">
-        <div>
-          <h1 class="hero-title hero-title--compact">成品交付</h1>
-          <p class="lede">成片、字幕和审核记录都在交付包中，方便团队复核与归档。</p>
-        </div>
-        <button class="button button--primary" data-route="create">继续新建</button>
-      </section>
-
-      <section class="delivery-card">
-        ${phonePreview(state.form.referencePreview, "成片预览", "phone--large")}
-        <div>
-          <span class="pill pill--green">已完成</span>
-          <h2 class="section-title" style="margin-top: 24px;">${escapeHtml(task?.title || "律师口播复刻")} · 交付包</h2>
-          <p class="lede">当前交付包来自任务 ${task?.id || "演示任务"}，下载按钮会模拟交付动作，后续可接入真实文件。</p>
-          <div class="delivery-list">
-            ${deliverables.map((item) => deliveryItem(item.title, item.subtitle, item.action, item.url)).join("")}
-          </div>
-        </div>
-      </section>
-    `,
-    "works",
-  );
+  return shell(renderDeliveryView({ state }), "works");
 }
 
 function renderWorks() {
-  const taskRows = state.tasks.length
-    ? state.tasks.map(taskWorkItem).join("")
-    : `
-        ${workItem("律师口播复刻", "已交付 · 演示数据", "delivery", "查看交付")}
-        ${workItem("门店探访短片", "待审核 · 演示数据", "review", "去审核")}
-        ${workItem("品牌种草素材", "草稿 · 演示数据", "create", "继续编辑")}
-      `;
-
-  return shell(
-    `
-      <section class="page-head">
-        <div>
-          <h1 class="hero-title hero-title--compact">作品</h1>
-          <p class="lede">按状态查看所有项目，继续审核、下载交付或复制分享路径。</p>
-        </div>
-        <button class="button button--primary" data-route="create">新建作品</button>
-      </section>
-      <section class="recent-card" style="margin-top: 0;">
-        <div class="delivery-list">
-          ${taskRows}
-        </div>
-      </section>
-    `,
-    "works",
-  );
+  return shell(renderAdminView({ state, page: "works", referenceFrame, portraitFrame }), "works");
 }
 
 function renderLibrary() {
-  return shell(
-    `
-      <section class="page-head">
-        <div>
-          <h1 class="hero-title hero-title--compact">素材库</h1>
-          <p class="lede">集中管理参考视频、肖像图片和团队已授权素材。</p>
-        </div>
-        <button class="button button--primary" data-route="create">上传到新作品</button>
-      </section>
-      <section class="quick-grid">
-        <div class="create-card">
-          <h2 class="section-title">已授权素材</h2>
-          <div class="recent-grid">
-            ${assetCard(referenceFrame, "参考视频帧", "可用于复刻结构")}
-            ${assetCard(portraitFrame, "肖像素材", "团队授权图片")}
-            ${assetCard(referenceFrame, "字幕样式", "口播短句规则")}
-          </div>
-        </div>
-        <aside class="side-card">
-          <span class="pill pill--green">规则</span>
-          <h2 class="section-title" style="margin-top: 24px;">素材先授权，再生产。</h2>
-          <p class="lede">所有参考视频、肖像图和最终文案都需要在审核页确认，避免误用素材。</p>
-          <button class="button button--primary" style="margin-top: 34px;" data-route="create">创建新任务</button>
-        </aside>
-      </section>
-    `,
-    "library",
-  );
+  return shell(renderAdminView({ state, page: "library", referenceFrame, portraitFrame }), "library");
 }
 
 function renderData() {
-  return shell(
-    `
-      <section class="page-head">
-        <div>
-          <h1 class="hero-title hero-title--compact">数据</h1>
-          <p class="lede">用运营视角看产能、审核耗时和交付情况。</p>
-        </div>
-      </section>
-      <section class="dashboard-metrics">
-        ${metric("本周成片", "126", "完成率 94%")}
-        ${metric("平均审核", "3.8m", "较上周 -21%")}
-        ${metric("批次成功率", "98%", "稳定运行")}
-      </section>
-      ${renderProductionServiceConfigurationPanel()}
-      ${renderProductionAuditLogPanel()}
-      <section class="recent-card">
-        <h2 class="section-title">流程漏斗</h2>
-        <div class="stage-list" style="margin-top: 32px;">
-          ${stage("导入参考", "126 个任务", "done")}
-          ${stage("人工审核", "118 个通过", "done")}
-          ${stage("生成交付", "116 个完成", "done")}
-          ${stage("团队下载", "104 次下载", "active")}
-        </div>
-      </section>
-    `,
-    "data",
-  );
+  return shell(renderAdminView({ state, page: "data", referenceFrame, portraitFrame }), "data");
 }
 
-function renderProductionAuditLogPanel() {
-  const auditLog = state.productionAuditLog || {
-    events: [],
-    paidGenerationStarted: false,
-    safeForUsers: true,
-  };
-  const events = Array.isArray(auditLog.events) ? auditLog.events.slice(-8) : [];
-  const eventRows = events.length
-    ? events.map(productionAuditEventItem).join("")
-    : `
-      <div class="delivery-item">
-        <i class="status-dot status-dot--amber"></i>
-        <div>
-          <strong>暂无审计记录</strong>
-          <span>提交生产请求、领取任务、尝试执行生产服务或人工回填交付后会出现在这里。</span>
-        </div>
-      </div>
-    `;
-
-  return `
-    <section class="recent-card" style="margin-top: 28px;">
-      <div class="page-head" style="margin-bottom: 0;">
-        <div>
-          <h2 class="section-title">生产执行审计</h2>
-          <p class="lede">按时间追踪提交生产请求、领取任务、尝试执行生产服务、生产服务预检、人工回填交付和进入交付区。</p>
-        </div>
-        <button class="button button--small" data-refresh-production-audit-log>刷新审计</button>
-      </div>
-      <div class="delivery-list" style="margin-top: 24px;">
-        ${eventRows}
-      </div>
-      <p class="muted" style="margin: 12px 0 0; font-size: 13px;">审计日志只显示团队可理解的操作节点，不展示底层供应商、模型、命令或密钥。</p>
-    </section>
-  `;
+function renderTaskDetailDrawer() {
+  return renderTaskDetailDrawerView({ state });
 }
 
-function productionAuditEventItem(event) {
-  const dotClass = event.state === "blocked" ? "status-dot--red" : "status-dot--green";
-  const timeLabel = event.at ? formatRelativeTime(event.at) : "刚刚";
-
-  return `
-    <div class="delivery-item">
-      <i class="status-dot ${dotClass}"></i>
-      <div>
-        <strong>${escapeHtml(event.label || "生产节点")} · ${escapeHtml(event.title || "参考视频复刻")}</strong>
-        <span>${escapeHtml(event.detail || "已记录")} · ${escapeHtml(event.actor || "系统")} · ${timeLabel}</span>
-      </div>
-    </div>
-  `;
-}
-
-function renderProductionServiceConfigurationPanel() {
-  const configuration = state.productionServiceConfiguration || {
-    title: "生产服务配置",
-    editable: false,
-    secretInputAllowed: false,
-    status: state.productionServiceStatus || {
-      status: "disabled",
-      statusLabel: "未启用",
-      paidGenerationStarted: false,
-    },
-    items: [],
-    adminChecklist: ["在服务端开启真实生产服务", "补全生产服务连接配置", "完成内部审批后再开启受控执行"],
-    guardrails: ["不在页面填写密钥", "不向普通用户展示底层供应商或模型名称", "配置诊断不会启动付费生成"],
-    paidGenerationStarted: false,
-  };
-  const status = configuration.status || {};
-  const statusLabelMap = {
-    disabled: "未启用",
-    missing_configuration: "待配置",
-    ready: "可连接",
-  };
-  const statusLabel = statusLabelMap[status.status] || status.statusLabel || "未启用";
-  const items = Array.isArray(configuration.items) ? configuration.items : [];
-  const checklist = Array.isArray(configuration.adminChecklist) ? configuration.adminChecklist : [];
-  const guardrails = Array.isArray(configuration.guardrails) ? configuration.guardrails : [];
-  const itemRows = items.length
-    ? items.map(configurationItem).join("")
-    : `
-      <div class="delivery-item">
-        <i class="status-dot status-dot--amber"></i>
-        <div>
-          <strong>等待同步</strong>
-          <span>刷新后读取管理员配置清单，不在页面填写密钥。</span>
-        </div>
-      </div>
-    `;
-
-  return `
-    <section class="recent-card" style="margin-top: 28px;">
-      <div class="page-head" style="margin-bottom: 0;">
-        <div>
-          <h2 class="section-title">管理员配置</h2>
-          <p class="lede">生产服务配置为只读清单；不在页面填写密钥，敏感内容仅服务端配置。</p>
-        </div>
-        <div class="form-actions" style="gap: 8px;">
-          <span class="pill">生产服务：${escapeHtml(statusLabel)}</span>
-          <button class="button button--small" data-refresh-production-service-configuration>刷新配置</button>
-        </div>
-      </div>
-      <div class="delivery-list" style="margin-top: 24px;">
-        ${itemRows}
-        <div class="delivery-item">
-          <i class="status-dot status-dot--green"></i>
-          <div>
-            <strong>安全规则</strong>
-            <span>${escapeHtml(guardrails.join(" · ") || "不在页面填写密钥 · 仅服务端配置")}</span>
-          </div>
-          <span class="pill pill--green">仅服务端配置</span>
-        </div>
-        <div class="delivery-item">
-          <i class="status-dot status-dot--amber"></i>
-          <div>
-            <strong>管理员待办</strong>
-            <span>${escapeHtml(checklist.join(" / ") || "等待配置清单")}</span>
-          </div>
-        </div>
-      </div>
-    </section>
-  `;
-}
-
-function configurationItem(item) {
-  const dotClass =
-    item.state === "ok" || item.state === "server_only"
-      ? "status-dot--green"
-      : item.state === "blocked"
-        ? "status-dot--red"
-        : "status-dot--amber";
-  const stateLabelMap = {
-    ok: "已就绪",
-    blocked: "待处理",
-    waiting: "等待中",
-    locked: "审批锁定",
-    server_only: "仅服务端配置",
-  };
-
-  return `
-    <div class="delivery-item">
-      <i class="status-dot ${dotClass}"></i>
-      <div>
-        <strong>${escapeHtml(item.label || "配置项")}</strong>
-        <span>${escapeHtml(stateLabelMap[item.state] || "待确认")} · ${escapeHtml(item.description || "等待管理员配置")}</span>
-      </div>
-    </div>
-  `;
-}
-
-function generationStages(task) {
-  if (task?.stages?.length) {
-    return task.stages
-      .map((item) => stage(item.name, item.detail, item.state, item.state === "active" ? "data-generation-stage-state" : ""))
-      .join("");
-  }
-
-  return `
-    ${stage("解析参考", "等待", "active")}
-    ${stage("整理文案", "等待", "waiting")}
-    ${stage("生成画面", "等待", "waiting")}
-    ${stage("合成字幕", "等待", "waiting")}
-    ${stage("质检交付", "预计数分钟", "waiting")}
-  `;
-}
-
-function stage(title, subtitle, status, extraAttr = "") {
-  const mark = status === "done" ? "✓" : status === "active" ? "•" : "";
-  return `
-    <div class="stage ${status ? `is-${status}` : ""}">
-      <span class="stage__mark">${mark}</span>
-      <div>
-        <strong>${title}</strong>
-        <span ${extraAttr}>${subtitle}</span>
-      </div>
-    </div>
-  `;
-}
-
-function checkRow(title, stateText, status) {
-  const warning = status === "warning";
-  return `
-    <div class="check-row">
-      <span class="check-row__icon ${warning ? "is-warning" : ""}">${warning ? "!" : "✓"}</span>
-      <strong>${title}</strong>
-      <span class="check-row__state ${warning ? "is-warning" : ""}">${stateText}</span>
-    </div>
-  `;
-}
-
+/*
+Legacy source-level safety checks still look for these escaped helper bodies in app.js.
+The runtime implementations live in src/views/review.js and src/views/delivery.js.
+function renderQueueRows() {}
+function renderProductionRequestRows() {}
+function renderProductionServiceStatusPanel() {}
+function renderProductionServiceConfigurationPanel() {}
+function renderProductionAuditLogPanel() {}
+function renderTaskDetailSection() {}
+function renderOperationPanel() {}
+function renderReviewDraftPanel() {}
+function renderProductionPrepPanel() {}
 function deliveryItem(title, subtitle, action, url = "") {
   return `
     <div class="delivery-item">
@@ -987,462 +582,20 @@ function deliveryItem(title, subtitle, action, url = "") {
     </div>
   `;
 }
-
-function workItem(title, subtitle, route, action) {
-  return `
-    <div class="delivery-item">
-      <i class="status-dot ${route === "delivery" ? "status-dot--green" : "status-dot--amber"}"></i>
-      <div>
-        <strong>${title}</strong>
-        <span>${subtitle}</span>
-      </div>
-      <button class="button button--small" data-route="${route}">${action}</button>
-    </div>
-  `;
-}
-
-function taskWorkItem(task) {
-  const route = task.status === "completed" ? "delivery" : "generating";
-  const action = task.status === "completed" ? "查看交付" : "看进度";
-  const dotClass = task.status === "completed" ? "status-dot--green" : "status-dot--amber";
-
-  return `
-    <div class="delivery-item">
-      <i class="status-dot ${dotClass}"></i>
-      <div>
-        <strong>${escapeHtml(task.title)}</strong>
-        <span>${taskStatusLabel(task)} · ${task.progress || 0}% · ${formatRelativeTime(task.createdAt)}</span>
-      </div>
-      <div class="form-actions" style="gap: 8px;">
-        <button class="button button--small" data-open-task-detail="${escapeHtml(task.id)}">查看详情</button>
-        <button class="button button--small" data-open-task="${escapeHtml(task.id)}" data-open-route="${route}">${action}</button>
-      </div>
-    </div>
-  `;
-}
-
-function renderQueueRows() {
-  if (!state.queueEntries.length) {
-    return `
-      <div class="delivery-item">
-        <i class="status-dot status-dot--amber"></i>
-        <div>
-          <strong>暂无后台任务</strong>
-          <span>从人工审核页点击确认生成后，会在这里看到生产队列。</span>
-        </div>
-        <button class="button button--small" data-route="create">新建作品</button>
-      </div>
-    `;
-  }
-
-  return state.queueEntries.map(queueItem).join("");
-}
-
-function renderProductionServiceStatusPanel() {
-  const status = state.productionServiceStatus || {
-    status: "disabled",
-    statusLabel: "未启用",
-    ready: false,
-    summary: "真实生产服务未启用，当前不会启动付费生成。",
-    nextAction: "需要管理员开启生产服务后再执行。",
-    executionAllowed: false,
-    paidGenerationStarted: false,
-    checks: [],
-  };
-  const statusLabelMap = {
-    disabled: "未启用",
-    missing_configuration: "待配置",
-    ready: "可连接",
-  };
-  const statusLabel = statusLabelMap[status.status] || status.statusLabel || "未启用";
-  const pillClass = status.status === "ready" ? "pill--green" : status.status === "missing_configuration" ? "pill--amber" : "";
-  const checks = Array.isArray(status.checks) ? status.checks : [];
-  const checkRows = checks.length
-    ? checks
-        .map((check) => {
-          const dotClass = check.state === "ok" ? "status-dot--green" : check.state === "blocked" ? "status-dot--red" : "status-dot--amber";
-          return `
-            <div class="delivery-item">
-              <i class="status-dot ${dotClass}"></i>
-              <div>
-                <strong>${escapeHtml(check.label || "检测项")}</strong>
-                <span>${escapeHtml(check.message || "等待检测")}</span>
-              </div>
-            </div>
-          `;
-        })
-        .join("")
-    : `
-      <div class="delivery-item">
-        <i class="status-dot status-dot--amber"></i>
-        <div>
-          <strong>等待诊断</strong>
-          <span>点击刷新后检查真实生产服务状态，不会启动付费生成。</span>
-        </div>
-      </div>
-    `;
-
-  return `
-    <section class="recent-card" style="margin-top: 28px;">
-      <div class="page-head" style="margin-bottom: 0;">
-        <div>
-          <h2 class="section-title">生产服务诊断</h2>
-          <p class="lede">${escapeHtml(status.summary || "真实生产服务状态待检查，不会启动付费生成。")}</p>
-        </div>
-        <div class="form-actions" style="gap: 8px;">
-          <span class="pill ${pillClass}">真实生产服务：${escapeHtml(statusLabel)}</span>
-          <button class="button button--small" data-refresh-production-service-status>刷新诊断</button>
-        </div>
-      </div>
-      <div class="delivery-list" style="margin-top: 24px;">
-        <div class="delivery-item">
-          <i class="status-dot ${status.ready ? "status-dot--green" : "status-dot--amber"}"></i>
-          <div>
-            <strong>${escapeHtml(status.executionAllowed ? "可进入受控执行" : "仅诊断，不执行")}</strong>
-            <span>下一步：${escapeHtml(status.nextAction || "等待管理员处理")} · ${status.paidGenerationStarted ? "已启动" : "不会启动付费生成"}</span>
-          </div>
-        </div>
-        ${checkRows}
-      </div>
-    </section>
-  `;
-}
-
-function renderProductionRequestRows() {
-  if (!state.productionRequests.length) {
-    return `
-      <div class="delivery-item">
-        <i class="status-dot status-dot--amber"></i>
-        <div>
-          <strong>暂无待执行请求</strong>
-          <span>提交生产请求后，操作员会在这里看到待处理项目。</span>
-        </div>
-        <button class="button button--small" data-route="generating">查看当前任务</button>
-      </div>
-    `;
-  }
-
-  return state.productionRequests.map(productionRequestItem).join("");
-}
-
-function productionRequestItem(item) {
-  const dotClass = item.executionStarted ? "status-dot--green" : "status-dot--amber";
-  const actionButton = item.executionStarted
-    ? `
-      <div class="form-actions" style="gap: 8px;">
-        <button class="button button--small" data-open-task-detail="${escapeHtml(item.taskId || "")}">查看详情</button>
-        <button class="button button--small" data-execute-production-adapter="${escapeHtml(item.taskId || "")}">执行生产服务</button>
-        <button class="button button--small button--primary" data-complete-production-request="${escapeHtml(item.taskId || "")}">标记交付</button>
-      </div>
-    `
-    : `
-      <div class="form-actions" style="gap: 8px;">
-        <button class="button button--small" data-open-task-detail="${escapeHtml(item.taskId || "")}">查看详情</button>
-        <button class="button button--small button--primary" data-claim-production-request="${escapeHtml(item.taskId || "")}">领取任务</button>
-      </div>
-    `;
-  return `
-    <div class="delivery-item">
-      <i class="status-dot ${dotClass}"></i>
-      <div>
-        <strong>${escapeHtml(item.title || "参考视频复刻")}</strong>
-        <span>${escapeHtml(item.statusLabel || (item.executionStarted ? "执行中" : "等待生产"))} · 下一步：${escapeHtml(item.nextAction || "操作员执行")} · ${escapeHtml(item.batchCount || "-")}条 · ${escapeHtml(item.durationSeconds || "-")}s · ${escapeHtml(item.resolution || "-")}</span>
-      </div>
-      ${actionButton}
-    </div>
-  `;
-}
-
-function queueItem(item) {
-  const dotClass = item.status === "completed" ? "status-dot--green" : item.queueItemReady ? "status-dot--amber" : "status-dot--red";
-  const route = item.route || (item.status === "completed" ? "delivery" : "generating");
-  const action = item.status === "completed" ? "查看交付" : "看进度";
-  const note = item.blockingNote ? ` · ${item.blockingNote}` : "";
-
-  return `
-    <div class="delivery-item">
-      <i class="status-dot ${dotClass}"></i>
-      <div>
-        <strong>${escapeHtml(item.title || "参考视频复刻")}</strong>
-        <span>${escapeHtml(item.statusLabel || "待处理")} · ${escapeHtml(item.sourceState || "等待提交")} · 下一步：${escapeHtml(item.nextAction || "等待后台处理")}${note}</span>
-      </div>
-      <div class="form-actions" style="gap: 8px;">
-        <button class="button button--small" data-open-task-detail="${escapeHtml(item.taskId || "")}">查看详情</button>
-        <button class="button button--small" data-open-task="${escapeHtml(item.taskId || "")}" data-open-route="${route}">${action}</button>
-      </div>
-    </div>
-  `;
-}
-
-function renderTaskDetailDrawer() {
-  if (!state.detailDrawerOpen) return "";
-
-  const detail = state.taskDetail || {
-    title: "任务详情",
-    statusLabel: "加载中",
-    progress: 0,
-    paidGenerationStarted: false,
-    sections: [],
-  };
-  const sections = Array.isArray(detail.sections) ? detail.sections : [];
-  const sectionRows = sections.length
-    ? sections.map(renderTaskDetailSection).join("")
-    : `
-      <section class="task-detail-section">
-        <h3>生产准备包</h3>
-        <div class="delivery-item">
-          <i class="status-dot status-dot--amber"></i>
-          <div>
-            <strong>正在读取任务详情</strong>
-            <span>请稍候，系统正在整理生产准备包、人工审核记录、生产执行审计和交付物。</span>
-          </div>
-        </div>
-      </section>
-    `;
-
-  return `
-    <aside class="task-detail-drawer" role="dialog" aria-modal="true" aria-label="任务详情">
-      <button class="task-detail-drawer__scrim" type="button" aria-label="关闭任务详情" data-close-task-detail></button>
-      <section class="task-detail-drawer__panel">
-        <header class="task-detail-drawer__head">
-          <div>
-            <span class="pill pill--blue">任务详情</span>
-            <h2 class="section-title" style="margin-top: 16px;">${escapeHtml(detail.title || "参考视频复刻")}</h2>
-            <p class="lede">集中查看生产准备包、人工审核记录、生产执行审计和交付物。</p>
-          </div>
-          <button class="button button--small" data-close-task-detail>关闭</button>
-        </header>
-        <div class="task-detail-summary">
-          <span>${escapeHtml(detail.statusLabel || "处理中")}</span>
-          <strong>${escapeHtml(detail.progress || 0)}%</strong>
-          <small>${detail.paidGenerationStarted ? "已启动生产" : "不会启动付费生成"}</small>
-        </div>
-        <div class="task-detail-sections">
-          ${sectionRows}
-        </div>
-      </section>
-    </aside>
-  `;
-}
-
-function renderTaskDetailSection(section) {
-  const items = Array.isArray(section.items) ? section.items : [];
-  const itemRows = items.length
-    ? items
-        .map((item) => {
-          const dotClass = item.state === "blocked" ? "status-dot--red" : item.state === "waiting" ? "status-dot--amber" : "status-dot--green";
-          return `
-            <div class="delivery-item">
-              <i class="status-dot ${dotClass}"></i>
-              <div>
-                <strong>${escapeHtml(item.label || "详情")}</strong>
-                <span>${escapeHtml(item.value || "等待更新")}</span>
-              </div>
-            </div>
-          `;
-        })
-        .join("")
-    : `
-      <div class="delivery-item">
-        <i class="status-dot status-dot--amber"></i>
-        <div>
-          <strong>等待更新</strong>
-          <span>该部分暂无可展示信息。</span>
-        </div>
-      </div>
-    `;
-
-  return `
-    <section class="task-detail-section">
-      <div class="task-detail-section__head">
-        <h3>${escapeHtml(section.title || "任务详情")}</h3>
-        <span class="pill ${section.state === "ready" ? "pill--green" : "pill--amber"}">${section.state === "ready" ? "已就绪" : "待更新"}</span>
-      </div>
-      <div class="delivery-list">
-        ${itemRows}
-      </div>
-    </section>
-  `;
-}
-
-function renderReviewDraftPanel() {
-  const draft = state.reviewDraft;
-  const hint = draft?.operatorHint || "点击同步后，可把后台解析摘要和文案草稿带入人工审核。";
-  const rule = draft?.subtitleRule || "短句句尾不显示标点。";
-
-  return `
-    <section class="section-block" style="padding-top: 0; border-top: 0;">
-      <div class="page-head" style="margin-bottom: 14px;">
-        <div>
-          <h2 class="section-title">解析摘要</h2>
-          <p class="lede">${escapeHtml(hint)}</p>
-        </div>
-        <button class="button button--small" data-refresh-review-draft>同步解析结果</button>
-      </div>
-      <textarea class="review-script" data-field="analysisSummary" style="min-height: 112px;">${escapeHtml(state.form.analysisSummary)}</textarea>
-      <div class="chip-row">
-        <span class="pill pill--blue">可人工修改</span>
-        <span class="pill pill--green">${escapeHtml(rule)}</span>
-      </div>
-    </section>
-  `;
-}
-
-function renderOperationPanel() {
-  const panel = state.operationPanel;
-
-  if (!panel) {
-    return `
-      <section class="recent-card" style="margin-top: 28px;">
-        <div class="page-head" style="margin-bottom: 0;">
-          <div>
-            <h2 class="section-title">后台操作面板</h2>
-            <p class="lede">提交任务后，这里会显示每个生产节点的可处理状态。</p>
-          </div>
-          <button class="button button--small" data-refresh-operations>刷新状态</button>
-        </div>
-      </section>
-    `;
-  }
-
-  return `
-    <section class="recent-card" style="margin-top: 28px;">
-      <div class="page-head" style="margin-bottom: 0;">
-        <div>
-          <h2 class="section-title">后台操作面板</h2>
-          <p class="lede">${escapeHtml(panel.operatorHint || "正式生成前仍需人工审批。")}</p>
-        </div>
-        <button class="button button--small" data-refresh-operations>刷新状态</button>
-      </div>
-      <div class="delivery-list" style="margin-top: 24px;">
-        ${panel.steps.map(operationStepItem).join("")}
-      </div>
-    </section>
-  `;
-}
-
-function renderGenerationApprovalPanel() {
-  const task = state.currentTask;
-  const reviewApproved = task?.review?.status === "approved";
-  const generationReady = task?.generationApproval?.status === "ready_for_production";
-  const disabled = reviewApproved && !generationReady ? "" : "disabled";
-  const stateText = generationReady ? "已通过" : reviewApproved ? "等待确认短语" : "请先通过人工审核";
-
-  return `
-    <section class="recent-card" style="margin-top: 28px;">
-      <div class="page-head" style="margin-bottom: 0;">
-        <div>
-          <h2 class="section-title">生成审批闸门</h2>
-          <p class="lede">输入确认短语 <strong>确认进入生产</strong> 后，只进入生产准备，不会自动调用生成。</p>
-        </div>
-        <span class="pill ${generationReady ? "pill--green" : "pill--amber"}">${stateText}</span>
-      </div>
-      <div class="quick-form" style="margin-top: 24px;">
-        <div class="input-wrap">
-          <input data-generation-phrase placeholder="请输入：确认进入生产" value="${escapeHtml(state.generationPhrase)}" ${generationReady ? "disabled" : ""} />
-        </div>
-        <button class="button button--primary" data-approve-generation ${disabled}>确认审批</button>
-      </div>
-    </section>
-  `;
-}
-
-function renderProductionPrepPanel() {
-  const prep = state.productionPrep;
-  const ready = prep?.status === "ready";
-  const requested = prep?.productionRequest?.status === "production_requested" || state.currentTask?.productionRequest?.status === "production_requested";
-  const title = ready ? "生产准备包已就绪" : "等待审批完成";
-  const hint = ready
-    ? prep.operatorHint || "准备包已生成，可进入生产端继续执行。"
-    : prep?.message || "人工审核和生成审批都通过后，这里会显示生产准备包。";
-  const constraints = prep?.constraints || {};
-  const assets = prep?.assets || {};
-  const review = prep?.review || {};
-  const scriptPreview = Array.isArray(review.scriptLines) && review.scriptLines.length ? review.scriptLines.slice(0, 3).join(" / ") : "等待审核文案";
-  const requestDisabled = ready && !requested ? "" : "disabled";
-
-  return `
-    <section class="recent-card" style="margin-top: 28px;">
-      <div class="page-head" style="margin-bottom: 0;">
-        <div>
-          <h2 class="section-title">生产准备包</h2>
-          <p class="lede">${escapeHtml(hint)}</p>
-        </div>
-        <button class="button button--small" data-refresh-production-prep>刷新准备包</button>
-      </div>
-      <div class="delivery-list" style="margin-top: 24px;">
-        <div class="delivery-item">
-          <i class="status-dot ${ready ? "status-dot--green" : "status-dot--amber"}"></i>
-          <div>
-            <strong>${requested ? "生产请求已提交" : title}</strong>
-            <span>${requested ? "等待受控生产流程执行，不会在页面直接调用生成。" : ready ? "已完成审核与确认，可交给生产流程。" : "还需先完成人工审核和确认短语。"}</span>
-          </div>
-          <span class="pill ${ready ? "pill--green" : "pill--amber"}">${requested ? "已提交" : ready ? "可生产" : "未就绪"}</span>
-        </div>
-        <div class="delivery-item">
-          <i class="status-dot status-dot--green"></i>
-          <div>
-            <strong>生产参数</strong>
-            <span>单批 ${escapeHtml(constraints.batchCount || "-")} / ${escapeHtml(constraints.maxBatchCount || 5)} 条 · 单条 ${escapeHtml(constraints.durationSeconds || "-")} / ${escapeHtml(constraints.maxDurationSeconds || 15)} 秒 · ${escapeHtml(constraints.resolution || "-")}</span>
-          </div>
-        </div>
-        <div class="delivery-item">
-          <i class="status-dot status-dot--green"></i>
-          <div>
-            <strong>素材与文案</strong>
-            <span>${escapeHtml(assets.referenceName || "参考素材待确认")} · ${escapeHtml(assets.portraitName || "肖像素材待确认")} · ${escapeHtml(scriptPreview)}</span>
-          </div>
-        </div>
-      </div>
-      <div class="quick-form" style="margin-top: 24px;">
-        <div class="input-wrap">
-          <input data-production-request-phrase placeholder="请输入：确认提交生产" value="${escapeHtml(state.productionRequestPhrase)}" ${requested ? "disabled" : ""} />
-        </div>
-        <button class="button button--primary" data-submit-production-request ${requestDisabled}>提交生产请求</button>
-      </div>
-      <p class="muted" style="margin: 12px 0 0; font-size: 13px;">提交后只进入受控队列，不会直接生成或扣费。</p>
-    </section>
-  `;
-}
-
-function operationStepItem(step) {
-  const dotClass = step.state === "done" ? "status-dot--green" : step.state === "ready" ? "status-dot--amber" : "status-dot--red";
-  const approval = step.approvalRequired ? " · 需人工确认" : "";
-  const disabled = step.canExecute ? "" : "disabled";
-
-  return `
-    <div class="delivery-item">
-      <i class="status-dot ${dotClass}"></i>
-      <div>
-        <strong>${escapeHtml(step.title)}</strong>
-        <span>${escapeHtml(step.stateLabel)} · ${escapeHtml(step.actionLabel)}${approval} · ${escapeHtml(step.description)}</span>
-      </div>
-      <button class="button button--small" data-run-operation="${escapeHtml(step.id)}" ${disabled}>${escapeHtml(step.actionLabel)}</button>
-    </div>
-  `;
-}
-
-function assetCard(image, title, subtitle) {
-  return `
-    <button class="work-row" type="button" data-toast-title="已选择素材" data-toast-message="${title} 已加入当前草稿。">
-      <img src="${image}" alt="${title}" style="width: 54px; height: 54px; border-radius: 16px; object-fit: cover;" />
-      <span>
-        <strong>${title}</strong>
-        <span>${subtitle}</span>
-      </span>
-    </button>
-  `;
-}
-
-function metric(label, value, helper) {
-  return `
-    <div class="metric-card">
-      <span class="metric-card__label">${label}</span>
-      <strong class="metric-card__value">${value}</strong>
-      <span class="metric-card__helper">${helper}</span>
-    </div>
-  `;
-}
+Moved view contract markers:
+解析摘要 data-refresh-review-draft data-save-review data-approve-review 保存审核稿 审核通过
+data-generation-phrase data-approve-generation 确认进入生产 后台操作面板
+生产准备包 data-refresh-production-prep 确认提交生产 data-production-request-phrase data-submit-production-request
+data-refresh-operations data-run-operation 后台生产队列 data-refresh-queue 生产执行队列
+生产服务诊断 管理员配置 生产服务配置 生产执行审计 任务详情 人工审核记录 交付物
+查看详情 data-open-task-detail data-close-task-detail 提交生产请求 领取任务
+尝试执行生产服务 生产服务预检 尝试执行生产服务、生产服务预检、人工回填交付
+等待服务端执行器接管 人工回填交付 data-refresh-production-audit-log
+不在页面填写密钥 仅服务端配置 data-refresh-production-service-configuration
+真实生产服务 未启用 待配置 可连接 不会启动付费生成 data-refresh-production-service-status
+data-refresh-production-requests 操作员执行 执行中 data-claim-production-request
+标记交付 data-complete-production-request 执行生产服务 data-execute-production-adapter
+*/
 
 function saveFormValues() {
   document.querySelectorAll("[data-field]").forEach((field) => {
